% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_sn.R
\name{sim_dynamic_sn}
\alias{sim_dynamic_sn}
\title{Simulation of dynamic network}
\usage{
sim_dynamic_sn(N, gamma, k0, phi, n_infs = 0, beta = 1, psi = 1,
  sigma = 1, nu = 1, eta = 1, mu = 1, rho = 0, n_vax = 0,
  vax_strat = NA, t, max.iter = 1e+06, burn.in = t, record = FALSE,
  record_term = 1, record_lengths = FALSE)
}
\description{
Simulation of dynamic network
}
