% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_sn_infs.R
\name{sim_outbreak_static_sn}
\alias{sim_outbreak_static_sn}
\title{Simulation of outbreak in static network}
\usage{
sim_outbreak_static_sn(N = 10000, gamma = 1.8, k0 = 0.5, phi = N,
  n_infs = 1, beta = 1, psi = 1, sigma = 1, nu = 1, eta = 1,
  mu = 1, rho = 0, t = 1, max.iter = 1e+06, sn = 0)
}
\arguments{
\item{N}{network size}

\item{gamma}{exponent of power law distribution}

\item{k0}{parameter controlling proportion of unpartnered}

\item{phi}{parameter controlling relationship turnaround rate}

\item{n_infs}{number of initially infected individuals}

\item{beta}{rate of transmission per partnership}

\item{psi}{proportion of infections becoming symptomatic}

\item{sigma}{reciprocal of duration of incubation period}

\item{nu}{reciprocal of duration of carriage}

\item{eta}{rate of screening when asymptomatic}

\item{mu}{reciprocal of time to seeking treatment}

\item{rho}{reciprocal of time to recovery following treatment}

\item{t}{duration of outbreak simulation}

\item{max.iter}{maximum number of Gillespie iterations}

\item{sn}{sexual network, if pre-computed}
}
\description{
Simulation of outbreak in static network
}
