% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_sn.R
\name{sim_static_sn}
\alias{sim_static_sn}
\title{Simulation of static network}
\usage{
sim_static_sn(N = 10000, gamma = 1.8, k0 = 0.5, phi = 10000)
}
\arguments{
\item{N}{network size}

\item{gamma}{exponent of power law distribution}

\item{k0}{parameter controlling proportion of unpartnered}

\item{phi}{parameter controlling relationship turnaround rate}
}
\description{
Simulation of static network
}
