% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{simdynet}
\alias{simdynet}
\title{simdynet project}
\description{
This is the simdynet project.

The main functions of the package are:
\itemize{

\item sim_dynamic_sn
\item sim_static_sn
\item sim_outbreak_static_sn
}
}
\references{
Whittles et al, Manuscript in prepration.
}
\seealso{
https://github.com/lwhittles/simdynet
}
\author{
Lilith Whittles \email{l.whittles14@imperial.ac.uk} and Xavier Didelot \email{xavier.didelot@gmail.com}
}
