% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_sn_infs.R
\name{sim_outbreak_static_sn}
\alias{sim_outbreak_static_sn}
\title{Simulation of outbreak in static network}
\usage{
sim_outbreak_static_sn(N, sn = 0, gamma, k0, phi, n_infs = 0, beta = 1,
  psi = 1, sigma = 1, nu = 1, eta = 1, mu = 1, rho = 0, t, max.iter)
}
\description{
Simulation of outbreak in static network
}
